Add transaction functionality

+++ Models/Transaction.cs
@@ -12,7 +12,7 @@ namespace Expense_Tracker.Models
         public int CategoryId { get; set; }
         public Category? Category { get; set; }
 
-        [Range(1, int.MaxValue, ErrorMessage = "Amount should be greater than 0.")]
+        [Range(1, double.MaxValue, ErrorMessage = "Amount should be greater than 0.")]
         public int Amount { get; set; }
 
         [Column(TypeName = "nvarchar(75)")]
